+++ plugins/source/aws/resources/services/acm/certificates_mock_test.go
@@ -57,5 +57,5 @@ func buildACMCertificates(t *testing.T, ctrl *gomock.Controller) client.Services
 }
 
 func TestACMCertificates(t *testing.T) {
-	client.AwsMockTestHelper(t, AcmCertificates(), buildACMCertificates, client.TestOptions{})
+	client.AwsMockTestHelper(t, Certificates(), buildACMCertificates, client.TestOptions{})
 }