+++ website/next.config.js
@@ -1,0 +1,23 @@
+const withNextra = require("nextra")({
+  theme: "nextra-theme-docs",
+  themeConfig: "./theme.config.tsx",
+  unstable_flexsearch: true,
+  unstable_staticImage: true,
+});
+
+module.exports = withNextra({
+  reactStrictMode: true,
+  experimental: {
+    legacyBrowsers: false,
+    images: { allowFutureImage: true },
+  },
+  async redirects() {
+    return [
+      {
+        source: "/docs/changelog",
+        permanent: true,
+        destination: "https://github.com/cloudquery/cloudquery/releases",
+      },
+    ];
+  },
+});